feet: @Transactional 애노테이션 적용
